organise portfolioId endpoints
